Added clear button functionality to the search box
