+++ maryam/core/web/webSearchInterface/src/App.js
@@ -4,7 +4,7 @@ import Home from "./pages/Home";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 // import Settings from './pages/Settings';
 import SearchPage from "./pages/SearchPage";
-import Image from "./pages/Image"
+// import Image from "./pages/Image"
 
 function App() {
   return (
@@ -15,8 +15,6 @@ function App() {
           <Route path="/" element={<Home />}>
           </Route>
           <Route path="/search" element={<SearchPage />}>
-          </Route>
-          <Route path="/image" element={<Image />}>
           </Route>
         </Routes>
       </Router>
