+++ maryam/core/web/webSearchInterface/src/pages/Search.css
@@ -67,3 +67,4 @@
 .shortCut_hidden {
   display: none !important;
 }
+
