+++ maryam/core/web/webSearchInterface/src/skeletons/SkeletonElement.css
@@ -1,5 +1,4 @@
 .skeleton {
-  /* background: gray.900; */
   background: #ddd;
   margin: 10px 0;
   border-radius: 4px;
