+++ maryam/core/web/webSearchInterface/src/skeletons/SkeletonSearchPage.js
@@ -1,12 +1,12 @@
 import Shimmer from "./Shimmer";
 import SkeletonElement from "./SkeletonElement";
 
-const SkeletonSearchPage = ({ theme }) =>{
+const SkeletonSearchPage = ({ theme, isScrolled}) =>{
     const themeClass= theme || 'light';
 
     return (
         <div className={`skeleton-wrapper ${themeClass}`}>
-            <div className="skeleton-searchPage">
+            <div className={`skeleton-searchPage ${isScrolled ? "headerOffset" : ""}`}>
                 <SkeletonElement type = "title"/>
                 <SkeletonElement type = "text"/>
                 <SkeletonElement type = "description"/>
